// frontend/src/app/interfaces/ToDoSecond.ts
export interface ToDoSecondInterface {
  id: string;
  userId: string;
  title: string;
  description: string;
  checkBoxList: any[];
  startTime: string;
  endTime: string;
  estimatedTime: number;
  progress: number;
  priority: number;
  desireToDo: number;
  avodiance: number;
  status: string;
}